chore(cli): trying to fix some bugs and cleaning
