+++ src/lib/latex/highlight.rs
@@ -108,58 +108,3 @@ where
         }
     }
 }
-
-#[cfg(feature = "cli")]
-pub(crate) mod cli {
-    use super::*;
-    use crate::cli::io::{InputArgs, OutputArgs};
-    use crate::cli::traits::Execute;
-    use clap::{Parser, ValueEnum};
-    use logos::Logos;
-
-    #[derive(Clone, Debug, ValueEnum)]
-    enum HighlightedPart {
-        Math,
-        Preamble,
-        Document,
-        InlineMath,
-        DisplayMath,
-    }
-
-    /// Highlight parts of TeX document(s) in a given color or return span locations.
-    #[derive(Debug, Parser)]
-    #[command(allow_missing_positional(true))]
-    pub struct HighlightCommand {
-        /// Part to be highlighted.
-        /// Required unless `--token` is present, in which case it will be ignored.
-        #[arg(required_unless_present("token"), value_enum)]
-        part: HighlightedPart,
-        /// Token to be highlighted.
-        #[arg(short, long, value_enum)]
-        pub token: Option<TokenDiscriminants>,
-        #[command(flatten)]
-        pub input_args: InputArgs,
-        #[command(flatten)]
-        pub output_args: OutputArgs,
-    }
-
-    impl Execute for HighlightCommand {
-        type Error = std::io::Error;
-        fn execute(self) -> Result<(), Self::Error> {
-            let mut stdout = self.output_args.stdout();
-            let sources = self.input_args.read_sources().unwrap();
-
-            let token = self.token.unwrap();
-            let color = self.output_args.color_args.into();
-
-            for source in sources.iter() {
-                let iter = Token::lexer(source.as_str()).spanned();
-
-                TokenHighlighter::new(iter, token)
-                    .write_colorized(source.as_str(), &mut stdout, &color)
-                    .unwrap();
-            }
-            Ok(())
-        }
-    }
-}
